django_app/requirements.txt
Django==1.10.3
argparse==1.2.1
django-chartit==0.2.7
django-chartjs==1.1
django-jquery==3.1.0
six==1.10.0
wsgiref==0.1.2
